+++ folio-kafka-wrapper/src/main/java/org/folio/kafka/KafkaProducerManager.java
@@ -1,7 +1,0 @@
-package org.folio.kafka;
-
-import io.vertx.kafka.client.producer.KafkaProducer;
-
-public interface KafkaProducerManager {
-  <K, V> KafkaProducer<K, V> createShared(String name);
-}